requirements.txt
alembic==1.14.0
bcrypt==4.2.0
blinker==1.9.0
boto3==1.35.68
botocore==1.35.68
cachelib==0.13.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
lxml==5.3.0
Mako==1.3.6
MarkupSafe==3.0.2
msgspec==0.18.6
pillow==11.0.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
redis==5.2.0
s3transfer==0.10.4
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
WTForms==3.2.1
gunicorn==23.0.0